Refactor fow layers with common fow super class
