+++ client/src/game/layers/fowvision.ts
@@ -1,36 +1,13 @@
-import { Layer } from "@/game/layers/layer";
 import { layerManager } from "@/game/layers/manager";
 import { gameStore } from "@/game/store";
 import { g2l, g2lr, g2lx, g2ly } from "@/game/units";
 import { gameSettingsStore } from "../settings";
 import { TriangulationTarget } from "../visibility/te/pa";
 import { computeVisibility } from "../visibility/te/te";
+import { FowLayer } from "./fow";
 import { floorStore } from "./store";
 
-export class FOWPlayersLayer extends Layer {
-    isVisionLayer = true;
-
-    virtualCanvas: HTMLCanvasElement;
-    vCtx: CanvasRenderingContext2D;
-
-    constructor(canvas: HTMLCanvasElement, public name: string, public floor: number, public index: number) {
-        super(canvas, name, floor, index);
-        this.virtualCanvas = document.createElement("canvas");
-        this.virtualCanvas.width = window.innerWidth;
-        this.virtualCanvas.height = window.innerHeight;
-        this.vCtx = this.virtualCanvas.getContext("2d")!;
-    }
-
-    set width(width: number) {
-        super.width = width;
-        this.virtualCanvas.width = width;
-    }
-
-    set height(height: number) {
-        super.height = height;
-        this.virtualCanvas.height = height;
-    }
-
+export class FowVisionLayer extends FowLayer {
     draw(): void {
         if (!this.valid) {
             // console.time("VI");
